AI (team): dynamically compute formation space. The formation distance will dynamically increase/decrease depending on how many other teammates are inside a specific range!
NOTE: We do count teammates (not enemies) in a specific range, even if they are NOT 'visible'.
This needs further explanation: Normally it is not good to 'see' other players through walls, because nobody can do that. In 'BotVisibleTeamMatesAndEnemies' it makes sense to exclude players behind walls because match influencing things are depending on this (damage teammates and enemies by kamikaze blast, etc.).
Within our new function 'BotCountTeamMates' it doesn't give bots any benefit during battle, neither against humans nor against other bots. So, being aware where the teammates are in this particular situation is okay, even if they are behind walls.
Someone will argue that this is not possible in real life. Anyways, it saves us a lot of work by doing it this way. Moreover we can argue against that, because there is some basic awareness of where teammates are even if they aren't visible (e.g.: chat/location).