Q3 AI FIX: Gap checking Nr.1. 'BotGapDistance' causes a few bugs: 1. Bots do slow down speed if there is a gap. 2. Although they slow down speed, they are told to jump (which is very bad). 3. Worst thing: although they are told to jump, they did NOT(!) check for gaps. 4. Indirectly related: 'BotGapDistance' is checked when it is not really needed (Travel_Walk), this causes bots to reduce speed in front of a jump reachability. In theory I think 'BotGapDistance' should only check for gaps in 'ravel_WalkOffLedge'. Why should a bot slow down speed if the gap is not really higher than he can jump etc.
This commit starts with fixing those bugs.
But at first, let's remove code that didn't work at all. Inside 'BotGapDistance' there is a check for when NOTHING is hit, this doesn't make sense, because on stairs the bot's trace WILL always hit solids. This check would make sense, if we check if the trace hits something, but this is nonsense, because this is always the case!
Anyways, this piece of code did nothing! You can test this with a simple print function!