Q3 AI FIX: correct contents Nr. 2.
In some cases CONTENTS_BOTCLIP was missing.
This will eventually fix some issues on some maps (swimming, prediction, etc.)

NOTE 1: Should we also use MASK_PLAYERSOLID in 'BotCheckBlocked'? Can we use this than to resolve the 'bots-standing-ontop-of-each-other-on-jumppads' issue?

NOTE 2: Correct the engine? Rename BOTCLIP to something more common? AICLIP? Avoid having BOTCLIP and MONSTERCLIP, because both are meaning the same.