AI (obstacle): Enhancement Nr. 1. This commit does nearly the same as this one https://github.com/zturtleman/mint-arena/commit/c71faf5c80b6be24fc9f7047c3b021e72e76d66f did. It will bots attract even more to go for the items they wanted to grab, before they had to enable the relevant areas to get the item.
NOTE: Checking for NBG while activating buttons etc. causes the bot to shoot or touch the button, but immediatly after this was done, he looks for another item (nearby). In most cases the item that was sealed and unreachable is an LTG not an NBG. This is the reason why the bot activates the obstacle and decide to grab the next NBG and don't enter the area where the LTG (he was hunting for) resides (although the area now is accessible).

So I changed NBG to LTG, since most of the time the (new) LTG will be the same as the one before the bot had to open dorrs etc. This will attract the bot more to go for the item he already opened the door for. A better solution would be to keep the LTG from before opening doors (= no seek for any goal) during the whole 'active' process. I can't hink of any solution to do this without reach too many node switches, so let's try this one, assuming the LTG stays the same.
Seems to work well.
Eh, this is hard to translate, but I think we know what we mean, even in the future :)