AI (battle sense): more elaborated attack distances. This still does interfere/contradict 'BotAggression' and therefore 'Retreat' behaviour. Nevertheless, this is also the case in vanilla Q3.
NOTE: Check if the new 'Obelisk' disatance fixes the bug on some maps, where bots won't attack the obelisk because they are too near to the obelisk (and will hurt themselves), but can't go farther away because of a wall behind them.

NOTE: Didn't we made all the weapons without projectiles using a special attack routine/funtion in the past?